# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole =
  (typeof RegisterInputRole)[keyof typeof RegisterInputRole];

export const RegisterInputRole = {
  pengguna: "pengguna",
  mitra: "mitra",
  merchant: "merchant",
} as const;

export interface RegisterInput {
  name: string;
  email: string;
  password: string;
  role: RegisterInputRole;
}

export type LoginInputRole =
  (typeof LoginInputRole)[keyof typeof LoginInputRole];

export const LoginInputRole = {
  pengguna: "pengguna",
  mitra: "mitra",
  merchant: "merchant",
} as const;

export interface LoginInput {
  email: string;
  password: string;
  role: LoginInputRole;
}

export interface RegisterPenggunaInput {
  name: string;
  phone: string;
  email: string;
  password: string;
  confirmPassword: string;
  agreeTerms: boolean;
}

export interface VerifyOtpBody {
  phone: string;
  otp: string;
}

export interface ResendOtpBody {
  phone: string;
}

export interface SendOtpResponse {
  message: string;
  phone: string;
  /** OTP code (dev only, remove in production with SMS integration) */
  otpCode?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  pengguna: "pengguna",
  mitra: "mitra",
  merchant: "merchant",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
  role: UserRole;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}